moved activity over to ReportIncidentActivity.kt
