+++ app/src/main/java/com/ucmobiledevelopment/freeways/MainActivity.kt
@@ -66,20 +66,16 @@ class MainActivity : ComponentActivity() {
                 viewModel.user = user
                 viewModel.listenToIncidents()
             }
-
-            viewModel.fetchIncidents(2019, 2020, 40, 1)
-            val incidents by viewModel.incidents.observeAsState(initial = emptyList())
             FreeWaysTheme {
                 // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colors.background,
                 ) {
-                    IncidentInfo("Android", incidents)
+
                 }
 
                 val context = LocalContext.current
-
                 Box(Modifier.fillMaxWidth(), contentAlignment = Alignment.Center) {
                     Row(Modifier
                         .padding(5.dp),
@@ -102,252 +98,7 @@ class MainActivity : ComponentActivity() {
     }
 
 
-    @Composable
-    fun IncidentInfo(name: String, incidents : List<Incident> = ArrayList<Incident>()) {
-        var inStateName by remember { mutableStateOf("") }
-        var inCountyName by remember { mutableStateOf("") }
-        var inCityName by remember { mutableStateOf("") }
-        var inLatitude by remember { mutableStateOf("") }
-        var inLongitude by remember { mutableStateOf("") }
-        var inWay1 by remember { mutableStateOf("") }
-        var inWay2 by remember { mutableStateOf("") }
-        var inVehiclesInvolved by remember { mutableStateOf("") }
-        val context = LocalContext.current
 
-        Box(Modifier.fillMaxWidth(), contentAlignment = Alignment.Center) {
-            Column {
-                OutlinedTextField(
-                    value = inStateName,
-                    onValueChange = { inStateName = it },
-                    label = { Text(stringResource(R.string.stateName)) },
-                    modifier = Modifier.fillMaxWidth()
-                )
-                OutlinedTextField(
-                    value = inCountyName,
-                    onValueChange = { inCountyName = it },
-                    label = { Text(stringResource(R.string.countyName)) },
-                    modifier = Modifier.fillMaxWidth()
-                )
-                OutlinedTextField(
-                    value = inCityName,
-                    onValueChange = { inCityName = it },
-                    label = { Text(stringResource(R.string.cityName)) },
-                    modifier = Modifier.fillMaxWidth()
-                )
-                OutlinedTextField(
-                    value = inLatitude,
-                    onValueChange = { inLatitude = it },
-                    label = { Text(stringResource(R.string.latitude)) },
-                    modifier = Modifier.fillMaxWidth()
-                )
-                OutlinedTextField(
-                    value = inLongitude,
-                    onValueChange = { inLongitude = it },
-                    label = { Text(stringResource(R.string.longitude)) },
-                    modifier = Modifier.fillMaxWidth()
-                )
-                OutlinedTextField(
-                    value = inWay1,
-                    onValueChange = { inWay1 = it },
-                    label = { Text(stringResource(R.string.way1)) },
-                    modifier = Modifier.fillMaxWidth()
-                )
-                OutlinedTextField(
-                    value = inWay2,
-                    onValueChange = { inWay2 = it },
-                    label = { Text(stringResource(R.string.way2)) },
-                    modifier = Modifier.fillMaxWidth()
-                )
-                OutlinedTextField(
-                    value = inVehiclesInvolved,
-                    onValueChange = { inVehiclesInvolved = it },
-                    label = { Text(stringResource(R.string.vehiclesInvolved)) },
-                    modifier = Modifier.fillMaxWidth()
-                )
-                Box(Modifier.fillMaxWidth(), contentAlignment = Alignment.Center)
-                {
-                    Row() {
-                        Button(
-                            onClick = {
-                                var incidentInfo = Incident().apply {
-
-                                    stateId =  0
-                                    stateName = inStateName
-                                    countyId = 0
-                                    countyName = inCountyName
-                                    cityName = inCityName
-                                    latitude = inLatitude
-                                    longitude = inLongitude
-                                    way1 = inWay1
-                                    way2 = inWay2
-                                    vehiclesInvolved = 0
-
-                                }
-                                viewModel.saveIncident(incidentInfo)
-                                Toast.makeText(
-                                    context,
-                                    "$inCityName $inCountyName $inStateName $inLatitude $inLongitude $inWay1 $inWay2 $inVehiclesInvolved",
-                                    Toast.LENGTH_LONG
-                                ).show()
-                            }
-                        )
-
-                        {
-                            Text(text = "Save")
-                        }
-
-                        Button(
-                            onClick = {
-                                signIn()
-                            }
-                        ) {
-                            Text(text = "Logon")
-                        }
-                        Button(
-                            onClick = {
-                                takePhoto()
-                            }
-                        ) {
-                            Text(text = "Photo")
-                        }
-                }
-          //TO DO: Add a button to display the image          //AsyncImage(model = strUri, contentDescription= "Incident Image")
-                }
-
-            }
-
-        }
-    }
-
-    private fun takePhoto() {
-        if (hasCameraPermission() == PERMISSION_GRANTED && hasExternalStoragePermission() == PERMISSION_GRANTED){
-            // User has already granted permission for these activities. Toggle the camera!
-            invokeCamera()
-        }
-        else{
-            // User has not granted permissions, so we must request.
-            requestMultiplePermissionsLauncher.launch(arrayOf(
-                Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                Manifest.permission.CAMERA
-            ))
-        }
-    }
-
-    private val requestMultiplePermissionsLauncher = registerForActivityResult(ActivityResultContracts.RequestMultiplePermissions()) {
-        resultsMap ->
-        var permissionGranted = false
-        resultsMap.forEach {
-            if (it.value == true) {
-                permissionGranted = it.value
-            }
-            else {
-                permissionGranted = false
-                return@forEach
-            }
-        }
-        if (permissionGranted) {
-            invokeCamera()
-        }
-        else {
-            Toast.makeText(this, getString(R.string.cameraPermissionDenied), Toast.LENGTH_LONG).show()
-        }
-    }
-
-    private fun invokeCamera() {
-        val file = createImageFile()
-        try{
-            uri = FileProvider.getUriForFile(this, "com.ucmobiledevelopment.freeways.fileprovider", file)
-        }
-        catch (e: Exception) {
-            Log.e(TAG, "Error: ${e.message}")
-            var foo = e.message
-        }
-        getCameraImage.launch(uri)
-    }
-
-    private fun createImageFile() : File {
-        val timestamp = SimpleDateFormat("yyyyMMdd_HHmmss").format(Date())
-        val imageDirectory = getExternalFilesDir(Environment.DIRECTORY_PICTURES)
-        return File.createTempFile(
-            "Incident_${timestamp}",
-            ".jpg",
-            imageDirectory
-        ).apply {
-            currentImagePath = absolutePath
-        }
-    }
-
-    private val getCameraImage = registerForActivityResult(ActivityResultContracts.TakePicture()) {
-        success ->
-        if (success) {
-            Log.i(TAG, "Image Location: $uri")
-            strUri = uri.toString()
-            val photo = Photo(localUri = uri.toString())
-            viewModel.photos.add(photo)
-        }
-        else {
-            Log.e(TAG, "Image not saved. $uri")
-        }
-    }
-
-    private fun hasCameraPermission() = ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA)
-    private fun hasExternalStoragePermission() = ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE)
-
-
-
-
-
-    @Composable
-    fun Greeting(name: String) {
-        Text(text = "Hello $name!")
-    }
-
-
-    @Preview(showBackground = true)
-    @Composable
-    fun DefaultPreview() {
-        FreeWaysTheme {
-            Greeting("Android")
-        }
-    }
-
-    private fun signIn() {
-        val providers = arrayListOf(
-            AuthUI.IdpConfig.EmailBuilder().build(),
-            AuthUI.IdpConfig.GoogleBuilder().build()
-
-        )
-        val signInIntent = AuthUI.getInstance()
-            .createSignInIntentBuilder()
-            .setAvailableProviders(providers)
-            .build()
-
-        signInLauncher.launch(signInIntent)
-
-    }
-
-    private val signInLauncher = registerForActivityResult(
-        FirebaseAuthUIActivityResultContract()
-    ){
-        res -> this.signInResult(res)
-    }
-
-    private fun signInResult(result: FirebaseAuthUIAuthenticationResult) {
-        val response = result.idpResponse
-        if (result.resultCode == RESULT_OK) {
-            firebaseUser = FirebaseAuth.getInstance().currentUser
-            firebaseUser?.let {
-                val user = User(it.uid, it.displayName)
-                viewModel.user = user
-                viewModel.saveUser()
-                viewModel.listenToIncidents()
-
-            }
-
-        } else {
-            Log.e("MainActivity.kt", "Error logging in " + response?.error?.errorCode)
-        }
-    }
 }
